requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
pywin32==308
requests==2.32.3
urllib3==2.2.3
usb-monitor==1.21
WMI==1.5.1
paho-mqtt
